server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "bs4>=0.0.2",
    "celery[redis]>=5.4.0",
    "docx2txt>=0.8",
    "fastapi[all]>=0.128.0",
    "fastembed>=0.7.4",
    "flower>=2.0.1",
    "langchain-community>=0.4.1",
    "langchain-postgres>=0.0.16",
    "langchain-text-splitters>=1.1.0",
    "langchain[openai]>=1.2.7",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-postgres>=3.0.4",
    "loguru>=0.7.3",
    "minio>=7.2.20",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "pypdf>=5.0.0",
    "scalar-fastapi>=1.6.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
